Update index.html
Includes a form that lets users search for contacts by name, phone, or address. When the user types in a search term and clicks the "Search" button, the form sends the input to the "/search" URL. Below the form, there's a link that takes users to the "/view" page to see all contacts.